requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
et-xmlfile==1.1.0
frozenlist==1.4.0
idna==3.6
meraki==1.40.1
multidict==6.0.4
numpy==1.26.2
openpyxl==3.1.2
pandas==2.1.4
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
tzdata==2023.3
urllib3==2.1.0
yarl==1.9.4
